refactor: フォルダ構造をOpenAPIのスキーマの構造に沿ったものへリファクタ
