+++ backend/src/api/paths.rs
@@ -1,0 +1,2 @@
+pub mod problems;
+pub mod submissions;
